konsultasi/requirements.txt

google-generativeai